Add datamatrix-util.scad
New library file will contain "friend" functions that
are mostly for internal use in the data_matrix module
but could also be useful outside the module for certain
use-cases.

In particular, this library has some helper functions
for relating data sizes to other properties of the
symbol.

Move dm_pad and dm_ecc to util/datamatrix-util.scad
in preparation for automating their use in the
data_matrix module.