quick response: add padding function
I decided to change the compact nibbles function to leave
a final, hanging nibble intact so that it can be detected
and treated differently by subsequent parser stages.

However, on thinking about this I have come to the
conclusion that nibbles are probably the wrong abstraction
anyway. So I will probably have to rewrite this whole part
in the future.