cleanup: rename reed-solomon.scad
In looking at the changes for quick response, it seems that
converting this to a general-purpose algorithm would take a lot
of refactoring. Also, I speculate that it would reduce
performance quite a bit.

With the current algorithm, datamatrix and quick response would
only have about 17 lines of code in common. The bulk of the
files are the large, precomputed data tables and these are not
shared in common at all.

So, I have decided to rename the reed-solomon files and API
functions to be specific to datamatrix and will make new
versions for quick response.