quick response: fix padding bug
The recent changes to have variable bit-width fields has
left a condition where the special EOM handling in the
padding function was still expecting 4-bit nibbles even
though the compact function was starting to return other
variations of valid bits for the final byte.

Added test cases to check for this oversight.