Add unit test framework
Generates PNG images for preview and rendered versions of both top-down
and default viewports of each unit test .scad file.

Created a set of test cases for UPC-A and EAN-13.

I think the results/ images should usually be binary identical to the
expected/ images. If not, use an image diff tool like Beyond Compare
or simple visual inspection for comparison.