code39: Fix backwards compatibility with OpenSCAD 2015.03
OpenSCAD 2015.03 didn't have support for the let statement but did have
support for the let assignment operator.

Since the variables here don't depend on each other, they can be easily made
backwards-compatible.

Extreme backwards compatibility isn't a goal of this project, but there are
some LTS operating systems out there that are still packaging 2015.03 so it is
best to support it when easy.