UPC: update to GS1 size-specifications
There are different size specifications/recommendations for different
purposes, but these seem to be the normal "product" size as opposed to the
double-size for bulk shipping boxes.

X stride is still 0.33 mm, height is 22.85, which is slightly smaller than
before.
Additionally, EAN-13 has a recommended left side quiet zone of 11 modules
and right side quiet zone of 7 modules as opposed to a uniform 9 module
quiet zone.